IMPROVED: KillServer.sh, PROGRESS: added route to restore default setings. MSG: I will have backups of all 4 cfg files and whenthis route will be called, I will first delete the existing cfg files and then restore copies from the backups

+++ shellScripts/killServer.sh
@@ -1,3 +1,10 @@
 #! /bin/bash
-PID=$(ps -aux |grep "/bin/bash ./dummy.sh" |grep -v auto |grep S+ |sed -e's/  */ /g' |cut -d ' ' -f 2)
-kill -9 $PID
+PID=$(ps -aux |grep "/bin/bash ./shellScripts/RunServer.sh" |grep -v grep |sed -e's/  */ /g' |cut -d ' ' -f 2)
+# echo $PID
+if [ "$PID" = "" ]; then
+    exit 1 
+else
+    kill -9 $PID
+    exit 0
+fi
+
